Move disconnect handler closer to the data source it tracks
Issue: #75
